Reorganise to allow for code sharing
