+++ build.sbt
@@ -34,6 +34,7 @@ val jacksonScalaModule: String = "2.13.3"
 val simpleConfigurationVersion: String = "1.5.6"
 val googleOAuthClient: String = "1.33.3"
 val nettyVersion: String = "4.1.78.Final"
+val slf4jVersion: String = "1.7.36"
 
 val standardSettings = Seq[Setting[_]](
   resolvers ++= Seq(
@@ -54,6 +55,8 @@ val standardSettings = Seq[Setting[_]](
 lazy val commoneventconsumer = project
   .settings(Seq(
     libraryDependencies ++= Seq(
+      "org.slf4j" % "slf4j-api" % slf4jVersion,
+      "com.amazonaws" % "aws-java-sdk-athena" % awsSdkVersion,
       "com.typesafe.play" %% "play-json" % playJsonVersion,
       "org.specs2" %% "specs2-core" % specsVersion % "test",
       "com.fasterxml.jackson.core" % "jackson-databind" % jacksonDatabind,
@@ -258,7 +261,7 @@ def lambda(projectName: String, directoryName: String, mainClassName: Option[Str
     ),
     libraryDependencies ++= Seq(
       "com.amazonaws" % "aws-lambda-java-core" % "1.2.1",
-      "org.slf4j" % "slf4j-api" % "1.7.36",
+      "org.slf4j" % "slf4j-api" % slf4jVersion,
       "com.gu" %% "simple-configuration-core" % simpleConfigurationVersion,
       "com.gu" %% "simple-configuration-ssm" % simpleConfigurationVersion,
       "ch.qos.logback" % "logback-classic" % "1.2.11",
@@ -337,7 +340,6 @@ lazy val eventconsumer = lambda("eventconsumer", "eventconsumer", Some("com.gu.n
       libraryDependencies ++= Seq(
         "com.typesafe.play" %% "play-json" % playJsonVersion,
         "com.amazonaws" % "aws-java-sdk-dynamodb" % awsSdkVersion,
-        "com.amazonaws" % "aws-java-sdk-athena" % awsSdkVersion,
         "org.scala-lang.modules" %% "scala-java8-compat" % "0.9.1",
         "io.netty" % "netty-codec-http2" % nettyVersion
       ),
@@ -346,7 +348,7 @@ lazy val eventconsumer = lambda("eventconsumer", "eventconsumer", Some("com.gu.n
   })
 
 lazy val sloMonitor = lambda("slomonitor", "slomonitor", Some("com.gu.notifications.slos.SloMonitor"))
-  .dependsOn(commoneventconsumer)
+  .dependsOn(common)
   .settings({
     Seq(
       description := "Monitors SLO performance for breaking news notifications",
